[#issue18]: Improve maintainability of cheatdetector
by use of interface and reset cheat information at end of round,
so cheating detection is round-wise.
