+++ WizardServer/src/main/java/at/aau/ase/CheatDetector.java
@@ -1,59 +1,30 @@
 package at.aau.ase;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import at.aau.ase.libnetwork.androidnetworkwrapper.networking.game.basic_classes.Card;
-import at.aau.ase.libnetwork.androidnetworkwrapper.networking.game.basic_classes.Color;
-import at.aau.ase.libnetwork.androidnetworkwrapper.networking.game.basic_classes.Hand;
-import at.aau.ase.libnetwork.androidnetworkwrapper.networking.game.basic_classes.Player;
 
-public class CheatDetector {
+/**
+ * Interface for CheatDetectors
+ */
+public interface CheatDetector {
 
-    private Game game;
-    private Hand[] playerHands;
-    private Integer activePlayerIdx;
-    private Hand table;
-    private Color trump;
-    private List<Player> players;
-    private boolean[] isCheating; // for all players
+    /**
+     * Called every time a new card is put on the table. Updates the cheating status of the player,
+     * depending on the card played and the cards in this players hands and the current active color of the round.
+     * @param playedCard, Card that a player wants to put on the table
+     */
+    public void update(Card playedCard);
 
-    public CheatDetector(Game game) {
-        this.game = game;
-        this.playerHands = game.getPlayerHands();
-        this.activePlayerIdx = game.getActivePlayerIndex();
-        this.table = game.getTable();
-        this.trump = game.getTrump();
-        this.players = game.getPlayers();
-        this.isCheating = new boolean[this.players.size()];
-    }
+    /**
+     * Method used to check if a player is cheating.
+     * @param playerSuspectedToCheat, String of the players name.
+     * @return true if the player playerSuspectedToCheat is actually cheating
+     * @return false if the player playerSuspectedToCheat is actually NOT cheating
+     */
+    public boolean check(String playerSuspectedToCheat);
 
-    public void update(Card playedCard) {
-        Hand handToCheck = playerHands[activePlayerIdx];
-        ArrayList<Card> cardsOnTable = table.getCards();
-        Color activeColor = game.getActiveColor();
-
-        if (playedCard.getColor() != activeColor) { // Possibly cheating, check players cardsInHand
-            for (Card cardInHand : handToCheck.getCards()) {
-                if (cardInHand.getColor() == activeColor) { // Cheating!
-                    isCheating[activePlayerIdx] = true;
-                    break;
-                }
-            }
-        }
-    }
-
-    public boolean check(String playerSuspectedToCheat) {
-        int playerToCheck = 0;
-        for (int i = 0; i < players.size(); i++) { // One of them will match
-            if (players.get(i).getName().equals(playerSuspectedToCheat)) {
-                playerToCheck = i;
-            }
-        }
-        return isCheating[playerToCheck];
-    }
-
-    public void reset() {
-        this.isCheating = new boolean[this.players.size()];
-    }
+    /**
+     * Resets the cheating states of all players to false, meaning that nobody is cheating.
+     * Should be called at end of round, if cheating is per-round.
+     */
+    public void reset();
 }
